requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
# flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict
torchdata
transformers
vllm
# vllm==0.8.4
wandb
packaging
uvicorn
fastapi
latex2sympy2_extended
math_verify
tensorboard
ujson
scikit-learn
mathruler
ninja